+++ src/app/report/page.tsx
@@ -2,7 +2,9 @@
 import Image from 'next/image'
 
 import Header from '@/components/common/header/Header'
-import Calendar from '@/components/Calendar'
+import Calendar from '@/components/report/Calendar'
+import TodayEcoStats from '@/components/report/TodayEcoStats'
+import WeeklyEcoChart from '@/components/report/WeeklyEcoChart'
 
 export default function page() {
   return (
@@ -12,8 +14,8 @@ export default function page() {
         <Header backOnClick={() => alert('onClick!')} title="탄소 기록함" />
         <Calendar />
         <div className="flex flex-col gap-[35px] w-100% h-100% py-12 px-5 bg-[#D3EDE8]">
-          <div className="w-full h-[142px] bg-beige rounded-[15px]">오늘 걸은 거리</div>
-          <div className="w-full h-[280px] bg-beige rounded-[15px]">통계</div>
+          <TodayEcoStats />
+          <WeeklyEcoChart />
         </div>
       </div>
     </div>
